requirements.txt
# ----------------------------
# Core LLM + API Clients
# ----------------------------
openai==1.40.1
google-generativeai==0.5.2     # Gemini API
# anthropic==0.25.8             # Claude API (uncomment if used)
requests==2.32.3
httpx==0.27.2

# ----------------------------
# Data Handling & Analysis
# ----------------------------
pandas==2.2.2
numpy==1.26.4

# ----------------------------
# Semantic Scoring & NLP
# ----------------------------
sentence-transformers==2.7.0   # For embeddings & similarity scoring
scikit-learn==1.5.1            # For cosine similarity, clustering, etc.

# ----------------------------
# Backend Framework & Server
# ----------------------------
fastapi==0.110.0
uvicorn==0.35.0
SQLAlchemy==2.0.34
alembic==1.13.2
psycopg[binary]==3.2.1

# ----------------------------
# Visualization & Dashboard (Optional)
# ----------------------------
streamlit==1.37.1
matplotlib==3.9.1
seaborn==0.13.2


# ----------------------------
# Reporting & Automation
# ----------------------------
gspread==6.1.2                 # Google Sheets integration (optional)
fpdf==1.7.2                     # PDF export
schedule==1.2.1                 # Lightweight task scheduler

# ----------------------------
# Environment Management
# ----------------------------
python-dotenv==1.0.1


#database and storage
SQLAlchemy==2.0.34
alembic==1.13.2
psycopg2-binary==2.9.9